requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
hdbscan>=0.8.33
scikit-learn>=1.3.0
openpyxl>=3.1.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
